requirements.txt
Django
mysqlclient
celery
uwsgi
tencentyun
requests
djangorestframework
BeautifulSoup
oauth2